Chapter 10: add std to some size_ts
(cherry picked from commit 90956d090a5b94359fe143d4a9c366cba155fc4a)

+++ Exercises/Modules/Chapter 10/Soln10_05.cpp
@@ -26,7 +26,7 @@ int main()
 template <typename T>
 T larger(T a, T b)
 {
-  static size_t counter{};
+  static unsigned counter{};
   std::cout << "This instantation has now been called " << ++counter << " time(s)\n";
   return a > b? a : b;
 }
